move API calls relating to pretix into their own file
