+++ apps/passport-server/src/services/pretixSync.ts
@@ -1,5 +1,12 @@
 import { DateRange } from "@pcd/passport-interface";
 import { PoolClient } from "pg";
+import {
+  fetchOrders,
+  fetchSubevents,
+  PretixConfig,
+  PretixOrder,
+  PretixSubevent,
+} from "../apis/pretixAPI";
 import { ParticipantRole, PretixParticipant } from "../database/models";
 import { fetchPretixParticipants } from "../database/queries/fetchPretixParticipants";
 import { insertParticipant } from "../database/queries/insertParticipant";
@@ -43,12 +50,16 @@ export function startPretixSync(context: ApplicationContext) {
   }
 }
 
+/**
+ * Synchronize Pretix state with Zupass state.
+ */
 async function sync(context: ApplicationContext, pretixConfig: PretixConfig) {
   const participants = await loadAllParticipants(pretixConfig);
   const participantEmails = new Set(participants.map((p) => p.email));
 
   console.log(
-    `[PRETIX] loaded ${participants.length} Pretix participants (visitors and residents) from Pretix, found ${participantEmails.size} unique emails`
+    `[PRETIX] loaded ${participants.length} Pretix participants (visitors and residents)` +
+      ` from Pretix, found ${participantEmails.size} unique emails`
   );
 
   const { dbPool } = context;
@@ -263,92 +274,4 @@ function deduplicateVisitorParticipants(
   }
 
   return Array.from(dedupedVisitors.values());
-}
-
-// Fetch all orders for a given event.
-async function fetchOrders(
-  pretixConfig: PretixConfig,
-  eventID: string
-): Promise<PretixOrder[]> {
-  const orders: PretixOrder[] = [];
-
-  // Fetch orders from paginated API
-  let url = `${pretixConfig.orgUrl}/events/${eventID}/orders/`;
-  while (url != null) {
-    console.log(`[PRETIX] Fetching ${url}`);
-    const res = await fetch(url, {
-      headers: { Authorization: `Token ${pretixConfig.token}` },
-    });
-    if (!res.ok) {
-      console.error(`Error fetching ${url}: ${res.status} ${res.statusText}`);
-      break;
-    }
-    const page = await res.json();
-    orders.push(...page.results);
-    url = page.next;
-  }
-
-  return orders;
-}
-
-// Fetch all item types for a given event.
-async function fetchSubevents(
-  pretixConfig: PretixConfig,
-  eventID: string
-): Promise<PretixSubevent[]> {
-  const orders: PretixSubevent[] = [];
-
-  // Fetch orders from paginated API
-  let url = `${pretixConfig.orgUrl}/events/${eventID}/subevents/`;
-  while (url != null) {
-    console.log(`[PRETIX] Fetching ${url}`);
-    const res = await fetch(url, {
-      headers: { Authorization: `Token ${pretixConfig.token}` },
-    });
-    if (!res.ok) {
-      console.error(`Error fetching ${url}: ${res.status} ${res.statusText}`);
-      break;
-    }
-    const page = await res.json();
-    orders.push(...page.results);
-    url = page.next;
-  }
-
-  return orders;
-}
-
-// A Pretix order. For our purposes, each order contains one ticket.
-interface PretixOrder {
-  code: string; // "Q0BHN"
-  status: string; // "p"
-  testmode: boolean;
-  secret: string;
-  email: string;
-  positions: PretixPosition[]; // should have exactly one
-}
-
-// Unclear why this is called a "position" rather than a ticket.
-interface PretixPosition {
-  id: number;
-  order: string; // "Q0BHN"
-  positionid: number;
-  item: number;
-  price: string;
-  attendee_name: string; // "Danilo Kim"
-  attendee_email: string;
-  subevent: number;
-}
-
-interface PretixSubevent {
-  id: number;
-  date_from?: string | null;
-  date_to?: string | null;
-}
-
-interface PretixConfig {
-  token: string;
-  orgUrl: string;
-  zuEventID: string;
-  zuVisitorEventID: string;
-  zuEventOrganizersItemID: number;
 }
